refactor: extract reified namespace plan ownership
